#12 - feat - Comment Controller, Service 기능 구현

+++ spring-board/src/main/java/com/project/springboard/service/CommentService.java
@@ -3,16 +3,21 @@ package com.project.springboard.service;
 import com.project.springboard.domain.Comment.Comment;
 import com.project.springboard.domain.Post.Post;
 import com.project.springboard.domain.User.User;
-import com.project.springboard.dto.CommentDTO;
+import com.project.springboard.dto.CommentCreateDTO;
+import com.project.springboard.dto.CommentDeleteDto;
+import com.project.springboard.dto.CommentUpdateDTO;
 import com.project.springboard.dto.UserRequestDto;
 import com.project.springboard.repository.CommentRepository;
 import com.project.springboard.repository.PostRepository;
 import com.project.springboard.repository.UserRepository;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class CommentService {
@@ -21,36 +26,31 @@ public class CommentService {
 	private final PostRepository postRepository;
 	private final CommentRepository commentRepository;
 
-	private Comment getExistingComment(final CommentDTO commentDto) {
-		return commentRepository.findById(commentDto.getId())
-			.orElseThrow(() -> new IllegalArgumentException("해당하는 Comment가 없습니다. id=" + commentDto.getId()));
+	private Comment getExistingComment(final CommentCreateDTO commentCreateDto) {
+		return commentRepository.findById(commentCreateDto.getId())
+			.orElseThrow(() -> new IllegalArgumentException("해당하는 Comment가 없습니다. id=" + commentCreateDto.getId()));
 	}
 
-	public CommentDTO saveComment(final CommentDTO commentDto) {
-		if (commentDto.getUser() == null) {
-			throw new IllegalArgumentException("User cannot be null");
-		}
-		if (commentDto.getPost() == null) {
+	public CommentCreateDTO saveComment(final User user, final CommentCreateDTO commentCreateDto) {
+		if (commentCreateDto.getPost() == null) {
 			throw new IllegalArgumentException("Post cannot be null");
 		}
 
-		User user = userRepository.findById(commentDto.getUser().getId())
-			.orElseThrow(() -> new IllegalArgumentException("User cannot be null"));
-		Post post = postRepository.findById(commentDto.getPost().getId())
+		Post post = postRepository.findById(commentCreateDto.getPost().getId())
 			.orElseThrow(() -> new IllegalArgumentException("Post cannot be null"));
 
 		Comment comment = Comment.builder()
 			.user(user)
 			.post(post)
-			.content(commentDto.getContent())
+			.content(commentCreateDto.getContent())
 			.build();
 
 		commentRepository.save(comment);
 
-		return CommentDTO.from(comment);
+		return CommentCreateDTO.from(comment);
 	}
 
-	public CommentDTO updateComment(final CommentDTO updatedDto) {
+	public CommentCreateDTO updateComment(final CommentCreateDTO updatedDto) {
 		Comment comment = getExistingComment(updatedDto);
 
 		Comment updatedComment = Comment.builder()
@@ -61,23 +61,45 @@ public class CommentService {
 			.build();
 		commentRepository.save(updatedComment);
 
-		return CommentDTO.from(updatedComment);
+		return CommentCreateDTO.from(updatedComment);
 	}
 
-	public void deleteComment(final CommentDTO commentDto) {
-		Comment comment = getExistingComment(commentDto);
-
-		commentRepository.delete(comment);
-	}
-
-	public List<CommentDTO> findAllByUser(final UserRequestDto userRequestDto) {
+	public List<CommentCreateDTO> findAllByUser(final UserRequestDto userRequestDto) {
 		User user = userRepository.findById(userRequestDto.getId())
 			.orElseThrow(() -> new IllegalArgumentException("User cannot find"));
 
 		List<Comment> comments = commentRepository.findAllByUser(user);
 
 		return comments.stream()
-			.map(CommentDTO::from)
+			.map(CommentCreateDTO::from)
 			.collect(Collectors.toList());
+	}
+
+	public CommentCreateDTO updateComment(final User user, final CommentUpdateDTO commentUpdateDto) {
+		Comment comment = commentRepository.findById(commentUpdateDto.getId())
+			.orElseThrow(() -> new IllegalArgumentException("Comment cannot find"));
+
+		if (!Objects.equals(user.getId(), comment.getUser().getId())) {
+			throw new IllegalArgumentException("User is not coincided");
+		}
+
+		Comment updatedComment = Comment.builder()
+			.id(comment.getId())
+			.user(comment.getUser())
+			.post(comment.getPost())
+			.content(commentUpdateDto.getContent())
+			.build();
+		commentRepository.save(updatedComment);
+		return CommentCreateDTO.from(updatedComment);
+	}
+
+	public void deleteComment(final User user, final CommentDeleteDto commentDeleteDto) {
+		Comment comment = commentRepository.findById(commentDeleteDto.getId())
+			.orElseThrow(() -> new IllegalArgumentException("Comment cannot find"));
+
+		if (!Objects.equals(user.getId(), comment.getUser().getId())) {
+			throw new IllegalArgumentException("User is not coincided");
+		}
+		commentRepository.delete(comment);
 	}
 }
